+++ src/controllers/tag.controller.js
@@ -5,8 +5,9 @@ import {isValidId} from "../utils/validators.js"
 
 export const index = async (req, res) => {
   try {
+    const userId = req.user.user ? req.user.user.id : req.user.id;  
     const [rows] = await pool.query
-    ('SELECT * FROM tags ORDER BY created_at DESC');
+    ('SELECT * FROM tags WHERE user_id = ? ORDER BY created_at DESC', [userId]);
     return res.status(200).json({
       tags: tagsListDecorator(rows)
     });
@@ -18,11 +19,12 @@ export const index = async (req, res) => {
 
 export const store = async (req, res) => {
   try {
-    const { name, user_id } = req.body;
+    const { name } = req.body;
+    const userId = req.user.user ? req.user.user.id : req.user.id;
 
-    if (!name || !user_id) {
+    if (!name) {
       return res.status(400).json({
-        message: "El nombre y el user_id son obligatorios"
+        message: "El nombre es obligatorio"
       });
     }
 
@@ -30,7 +32,7 @@ export const store = async (req, res) => {
 
    await pool.query(
       'INSERT INTO tags (id, name, user_id) VALUES (?, ?, ?)',
-      [id, name, user_id]
+      [id, name, userId]
     );
 
     const [rows] = await pool.query(
@@ -55,6 +57,7 @@ export const store = async (req, res) => {
 export const show = async (req, res) => {
   try {
     const { id } = req.params;
+    const userId = req.user.user ? req.user.user.id : req.user.id;
 
     if (!isValidId(id)) {
       return res
@@ -63,8 +66,8 @@ export const show = async (req, res) => {
     }
 
     const [rows] = await pool.query(
-      'SELECT * FROM tags WHERE user_id = ? ORDER BY name ASC',
-      [id]
+      'SELECT * FROM tags WHERE id = ? AND user_id = ? ORDER BY name ASC',
+      [id,userId]
     );
 
     return res.status(200).json({
@@ -85,6 +88,7 @@ export const update = async (req, res) => {
   try {
     const { id } = req.params;
     const { name } = req.body;
+    const userId = req.user.user ? req.user.user.id : req.user.id;
 
     if (!isValidId(id)) {
       return res.status(400).json({ message: 'Identificador de etiqueta no válido' });
@@ -95,7 +99,7 @@ export const update = async (req, res) => {
     }
 
     const [result] = await pool.query(
-      'UPDATE tags SET name = ? WHERE id = ?', [name, id]
+      'UPDATE tags SET name = ? WHERE id = ? AND user_id = ?', [name, id, userId]
     );
 
     if (result.affectedRows === 0) {
@@ -105,7 +109,7 @@ export const update = async (req, res) => {
     }
 
     const [rows] = await pool.query(
-      'SELECT * FROM tags WHERE id = ?', [id]
+      'SELECT * FROM tags WHERE id = ? AND user_id = ?', [id, userId]
     );
 
     return res.status(200).json({
@@ -126,14 +130,15 @@ export const update = async (req, res) => {
 export const destroy = async (req, res) => {
   try {
     const { id } = req.params;
+    const userId = req.user.user ? req.user.user.id : req.user.id;
 
     if (!isValidId(id)) {
       return res.status(400).json({ message: 'Identificador de etiqueta no válido' });
     }
 
     const [result] = await pool.query(
-      'DELETE FROM tags WHERE id = ?',
-      [id]
+      'DELETE FROM tags WHERE id = ? AND user_id = ?',
+      [id, userId]
     );
 
     if (result.affectedRows === 0) {
