+++ src/controllers/task.controller.js
@@ -14,7 +14,8 @@ const getTaskTags = async (taskId) => {
 
 export const index = async (req, res) => {
   try {
-    const [tasks] = await pool.query('SELECT * FROM tasks ORDER BY created_at DESC');
+    const userId = req.user.user ? req.user.user.id : req.user.id;
+    const [tasks] = await pool.query('SELECT * FROM tasks WHERE user_id = ? ORDER BY created_at DESC', [userId]);
 
     if (tasks.length === 0) {
       return res.status(200).json({ tasks: [] });
@@ -66,11 +67,11 @@ export const index = async (req, res) => {
 
 export const store = async (req, res) => {
   try {
-    const { title,  description, status, category_id, user_id, tags } = req.body;
-
-    if (!title || !category_id || !user_id) {
+    const { title,  description, status, category_id, tags } = req.body;
+    const userId = req.user.user ? req.user.user.id : req.user.id;
+    if (!title || !category_id) {
       return res.status(400).json({
-        message: "El titulo, categoria_id y el user_id son obligatorios"
+        message: "El titulo, categoria_id son obligatorios"
       });
     }
 
@@ -79,7 +80,7 @@ export const store = async (req, res) => {
 
    await pool.query(
       'INSERT INTO tasks (id, title, description, status, category_id, user_id) VALUES (?, ?, ?, ?, ?, ?)',
-      [newTaskId, title, description || null, taskStatus, category_id, user_id]
+      [newTaskId, title, description || null, taskStatus, category_id, userId]
     );
 
     if (Array.isArray(tags) && tags.length > 0) {
@@ -105,6 +106,7 @@ export const store = async (req, res) => {
 export const show = async (req, res) => {
   try {
     const { id } = req.params;
+    const userId = req.user.user ? req.user.user.id : req.user.id;
 
     if (!isValidId(id)) {
       return res
@@ -113,8 +115,8 @@ export const show = async (req, res) => {
     }
 
     const [rows] = await pool.query(
-      'SELECT * FROM tasks WHERE id = ? ORDER BY name ASC',
-      [id]
+      'SELECT * FROM tasks WHERE id = ? AND user_id = ? ORDER BY name ASC',
+      [id, userId]
     );
 
     const task = rows[0];
@@ -138,24 +140,24 @@ export const show = async (req, res) => {
 export const update = async (req, res) => {
   try {
     const { id } = req.params;
-    const { title, description, status, category_id, user_id } = req.body;
-
+    const { title, description, status, category_id, tags } = req.body;
+    const userId = req.user.user ? req.user.user.id : req.user.id;
     if (!isValidId(id)) {
       return res.status(400).json({ message: 'Identificador de tarea no válido' });
     }
 
-    const [existing] = await pool.query('SELECT * FROM tasks WHERE id = ?', [id]);
+    const [existing] = await pool.query('SELECT * FROM tasks WHERE id = ? AND user_id = ?', [id, userId]);
     if (existing.length === 0) {
       return res.status(404).json({ message: 'Tarea no encontrada' });
     }
 
-    if (!title || !category_id || !user_id) {
-      return res.status(400).json({ message: 'Título, categoría y usuario requeridos' });
+    if (!title || !category_id) {
+      return res.status(400).json({ message: 'Título y categoría requeridos' });
     }
 
     await pool.query(
-      'UPDATE tasks SET title = ?, description = ?, status = ?, category_id = ?, user_id = ? WHERE id = ?',
-      [title, description || null, status || existing[0].status, category_id, user_id, id]
+      'UPDATE tasks SET title = ?, description = ?, status = ?, category_id = ?, user_id = ? WHERE id = ? and user_id = ?',
+      [title, description || null, status || existing[0].status, category_id, userId, id, userId]
     );
 
      if (Array.isArray(tags)) {
@@ -181,12 +183,13 @@ export const update = async (req, res) => {
 export const destroy = async (req, res) => {
    try {
     const { id } = req.params;
+    const userId = req.user.user ? req.user.user.id : req.user.id;
 
     if (!isValidId(id)) {
       return res.status(400).json({ message: 'Identificador de tarea no válido' });
     }
 
-    const [existing] = await pool.query('SELECT * FROM tasks WHERE id = ?', [id]);
+    const [existing] = await pool.query('SELECT * FROM tasks WHERE id = ? AND user_id = ?', [id, userId]);
     if (existing.length === 0) {
       return res.status(404).json({ message: 'Tarea no encontrada' });
     }
@@ -194,7 +197,7 @@ export const destroy = async (req, res) => {
     const [categoryRows] = await pool.query('SELECT * FROM categories WHERE id = ?', [task.category_id]);
     const tags = await getTaskTags(id);
     const deletedTaskData = taskDecorator(task, categoryRows[0] || null, tags);
-    await pool.query('DELETE FROM tasks WHERE id = ?', [id]);
+    await pool.query('DELETE FROM tasks WHERE id = ? AND user_id = ?', [id, userId]);
 
     return res.status(200).json({
       task: deletedTaskData
